# Speedup protocol checking