--- a/changes/2251.optimization.md
+++ b/changes/2251.optimization.md
@@ -1,0 +1,1 @@
+Improve protocol checking speed.