Using: CPython 3.13

Before
```
$ python scripts/benchmarks/protocol_benchmark.py
isinstance(long, BasicPyProtocol) 0.2666777889971854 µs
isinstance(long, BasicHikariProtocol) 0.6734864840036607 µs
isinstance(failfast, BasicPyProtocol) 2.1686498939961893 µs
isinstance(failfast, BasicHikariProtocol) 0.1939030650028144 µs
issubclass(long, BasicPyProtocol) 0.3827833930045017 µs
issubclass(long, BasicHikariProtocol) 0.4775878080035909 µs
issubclass(failfast, BasicPyProtocol) 0.3736991219993797 µs
issubclass(failfast, BasicHikariProtocol) 0.4793318279989762 µs
```

After:
```
$ python scripts/benchmarks/protocol_benchmark.py
isinstance(long, BasicPyProtocol) 0.27424019099998986 µs
isinstance(long, BasicHikariProtocol) 0.2070195690030232 µs
isinstance(failfast, BasicPyProtocol) 2.203423020997434 µs
isinstance(failfast, BasicHikariProtocol) 0.2940910609977436 µs
issubclass(long, BasicPyProtocol) 0.3797524759938824 µs
issubclass(long, BasicHikariProtocol) 0.17503146999661112 µs
issubclass(failfast, BasicPyProtocol) 0.39065740699879825 µs
issubclass(failfast, BasicHikariProtocol) 0.18609409200144 µs
```